Tentative success on adequate measure arithmetic.
